requirements.txt
torch>=2.8.0
transformers>=4.56.1
qdrant-client>=1.15.1
paddlepaddle>=3.1.0
paddleocr>=3.2.0
fastapi>=0.116.1
uvicorn>=0.35.0
pydantic>=2.11.7
numpy>=1.24.0
pillow>=10.0.0
pyyaml>=6.0
onnxruntime>=1.16.0
watchdog>=3.0.0
pytest>=7.4.0
sqlite-fts4